% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Send Some Praise}
\usage{
praise(name = "Roger")
}
\arguments{
\item{name}{character, the name of the person you want to praise}
}
\description{
Praise someone at critical times in my life
}
\examples{
## This is how you use it
praise()
}
