% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{praise_parts}
\alias{praise_parts}
\title{Parts of speech for praising}
\format{Named list of character vertors. List entries: \describe{
  \item{adjective}{Words and phrases to be used as positive adjectives.
    Most of them are from \url{https://github.com/sindresorhus/superb}.}
  \item{adverb}{Adverbs.}
  \item{adverb_manner}{Adverbs of manner, with positive meanings.}
  \item{created}{Synonyms of \sQuote{create} in paste tense.}
  \item{creating}{Synonyms of \sQuote{create}, in present participle
    form.}
  \item{exclamation}{Positive exclamations.}
  \item{rpackage}{Synonyms for the term \sQuote{R package}.}
}}
\usage{
praise_parts
}
\description{
Parts of speech for praising
}
\keyword{datasets}
