% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{praise_parts_langs}
\alias{praise_parts_langs}
\title{Parts of speech for praising, in all supported languages.}
\format{It is a named list, and each element is a language. The names of the
list elements are lowercase two letter ISO 639-1 codes.

Each list element is a list itself, and contains different parts of
speech. E.g. for English we have: \describe{
  \item{adjective}{Words and phrases to be used as positive adjectives.
    Most of them are from \url{https://github.com/sindresorhus/superb}.}
  \item{adverb}{Adverbs.}
  \item{adverb_manner}{Adverbs of manner, with positive meanings.}
  \item{created}{Synonyms of \sQuote{create} in paste tense.}
  \item{creating}{Synonyms of \sQuote{create}, in present participle
    form.}
  \item{exclamation}{Positive exclamations.}
  \item{rpackage}{Synonyms for the term \sQuote{R package}.}
}}
\usage{
praise_parts_langs
}
\description{
Parts of speech for praising, in all supported languages.
}
\keyword{datasets}
