% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pre_text}
\alias{pre_text}
\title{Randomized praise based on a template}
\format{
An object of class \code{character} of length 1.
}
\usage{
pre_text
}
\arguments{
\item{template}{Character scalar, the template string.}
}
\description{
Randomized praise based on a template
}
\details{
Replace parts of the template with random words from the praise
word lists. See examples below.
}
\examples{
praise()

## Capitalization
praise("${Exclamation}! This ${rpackage} is ${adjective}!")

## All upper case
praise("${EXCLAMATION}! You have done this ${adverb_manner}!")
}
\keyword{datasets}
