% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{superb}
\alias{superb}
\title{Words and phrases to be used as positive adjectives}
\format{A character vector of words and expressions.}
\usage{
superb
}
\description{
Words and phrases to be used as positive adjectives
}
\details{
Most of them are from \url{https://github.com/sindresorhus/superb}.
}
\keyword{datasets}

