% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Praise Someone}
\usage{
praise(name = "Holly", adjective = "best", punctuation = "!")
}
\arguments{
\item{name}{Specifies recipient of praise}

\item{adjective}{type of praise phrase you want to send}

\item{punctuation}{sets the tone of the praise}
}
\description{
Let's dole out praise
}
\examples{
praise(name = "Banana",
adjective = "best"
punctuation = "!?")

}
