% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Send some praise}
\usage{
praise(name = "Kim")
}
\arguments{
\item{name}{character, the name of the person you want to praise}
}
\description{
Praise someone at critical times in their life
}
\examples{
# I'm going to deliver some praise
praise()

}
