% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{praise}
\alias{praise}
\alias{praise-package}
\title{Praise Users}
\description{
Build friendly R packages that
praise their users if they have done something
good, or they just need it to feel better.
}

